Rename theme-app-extension-next to theme-app-extension
